+++ config/bootstrap.php
@@ -22,16 +22,14 @@ use MeCms\Database\Type\JsonEntityType;
 require_once __DIR__ . DS . 'constants.php';
 
 //Sets files to be copied
-Configure::write('CONFIG_FILES', [
-    ...Configure::read('CONFIG_FILES', []),
+Configure::write('MeCms.ConfigFiles', [
     'MeCms.recaptcha',
     'MeCms.me_cms',
     'MeCms.widgets',
 ]);
 
 //Sets directories to be created and must be writable
-Configure::write('WRITABLE_DIRS', [
-    ...Configure::read('WRITABLE_DIRS', []),
+Configure::write('MeCms.WritableDirs', [
     getConfigOrFail('Assets.target'),
     THUMBER_TARGET,
     UPLOADED,
@@ -40,10 +38,10 @@ Configure::write('WRITABLE_DIRS', [
 ]);
 
 //Sets symbolic links for vendor assets to be created
-Configure::write('VENDOR_LINKS', [
-    'studio-42' . DS . 'elfinder' => 'elfinder',
-    'enyo' . DS . 'dropzone' . DS . 'dist' => 'dropzone',
-] + Configure::read('VENDOR_LINKS', []));
+Configure::write('MeCms.VendorLinks', [
+    'studio-42/elfinder' => 'elfinder',
+    'enyo/dropzone/dist' => 'dropzone',
+]);
 
 //Sets configuration for the Tokens plugin
 Configure::write('Tokens.usersClassOptions', [