+++ tests/TestCase/Command/Install/SetPermissionsCommandTest.php
@@ -16,7 +16,7 @@ declare(strict_types=1);
 
 namespace MeCms\Test\TestCase\Command\Install;
 
-use Cake\Core\Configure;
+use MeTools\Core\Configure;
 use MeTools\TestSuite\CommandTestCase;
 use Tools\Filesystem;
 
@@ -33,7 +33,8 @@ class SetPermissionsCommandTest extends CommandTestCase
     {
         $this->exec('me_cms.set_permissions -v');
         $this->assertExitSuccess();
-        $expected = array_map(fn(string $path): string => 'Set permissions on `' . Filesystem::instance()->rtr($path) . '`', array_clean(Configure::read('WRITABLE_DIRS')));
-        $this->assertSame($expected, $this->_out->messages());
+        foreach (Configure::read('MeCms.WritableDirs') as $expectedDir) {
+            $this->assertOutputContains('Set permissions on `' . Filesystem::instance()->rtr($expectedDir) . '`');
+        }
     }
 }